Initial commit.
Simply ran `rails plugin new art_vandelay --mountable` to get a plugin
scaffold. It's possible that we do not need all the generated files, but
those can be removed in a future commit.

The plugin was generated with rails v7.0.4, but the version was removed
from the Gemfile in an effort to allow this plugin to work with older
version of Rails.