Add CI (#8)
This commit adds a CI workflow through GitHub actions, and is inspired
by [capybara_accessibility_audit].

Because CI runs against a matrix of Ruby versions and Rails versions,
some special considerations needed to be made.

- The Gemfile.lock was removed and ignored to avoid noisy diffs when
  working locally.
- Logic was added to the Gemfile to load the desired version of Rails,
  and account for some bugs with Rails 6 and certain versions of Ruby.
- Conditional logic was placed in a few files (migrations, application
  record) and the schema was updated to account for differences between
  versions

A future commit will introduce a setup script and set the default
version of Ruby.

[capybara_accessibility_audit]: https://github.com/thoughtbot/capybara_accessibility_audit/blob/main/.github/workflows/ci.yml