CI: Update Matrix and bump local Ruby version (#35)
Update Ruby Matrix to include supported, stable releases of Ruby, as
well as Rails `v7.1`. Drops support for Rails `v6.0`.

Updates local version of Ruby to the latest stable version, which at the
time of this writing is `3.3.4`.

Pins `sqlite3` to `~> 1.4` to avoid the following error when running
`bin/setup`:

```
Error loading the 'sqlite3' Active Record adapter
```

This is because Rails [supports][] only that version

[supports]: https://github.com/rails/rails/blob/v7.1.3.4/activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb#L14