Add `Import#json`, `Export#json`, and `Export#email` (#34)
This pull request allows the importing and exporting of JSON files. Previously only CSVs were supported. The interfaces for importing and exporting JSON have been designed to work the same way they already work for the CSV interfaces.

I've added documentation to the README for all of the JSON interfaces, so I won't go into more detail here. Refer to the changes to the README.

## Breaking changes

There is one breaking change here, which is that the `Export#email_csv` method has been renamed to `Export#email`. It takes a new `:format` argument which defaults to `:csv`. Users will need to update their application code accordingly:

```diff
-Export.email_csv(to: my_email)
+Export.email(to: my_email)
```